fix: Add discount markdown as alternative parser
The awk script is nice, but it doesn't parse codeblocks, and was a bit limited.

Although discount's markdown is written in C, it is still a bit bulky. Still, I would recommend it as a more convenient option. This represents a compromise between suckless principles and usability, so I have not set this as the default option.